Cleanup some of the internal API usage in Relational
- Make EnumerableExtensions shared internal
- Make ExpressionEqualityComparer public
- Suppress Internal usage in EntityFrameworkRelationalServicesBuilder
- Make ConfigurationSourceExtensions public
- Suppress convention-less modelBuilder creation in Migration/Snapshot
