+++ src/EFCore.Relational/Extensions/RelationalQueryableExtensions.cs
@@ -6,7 +6,6 @@ using System.Collections;
 using System.Data.Common;
 using System.Linq;
 using System.Linq.Expressions;
-using System.Reflection;
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore.Diagnostics;
 using Microsoft.EntityFrameworkCore.Query;
