+++ src/EFCore.Relational/Metadata/IColumnBase.cs
@@ -1,7 +1,6 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System.Collections;
 using System.Collections.Generic;
 using Microsoft.EntityFrameworkCore.Infrastructure;
