+++ src/EFCore.Relational/Metadata/Internal/TableIndexExtensions.cs
@@ -3,7 +3,7 @@
 
 using System.Text;
 using JetBrains.Annotations;
-using Microsoft.EntityFrameworkCore.Infrastructure.Internal;
+using Microsoft.EntityFrameworkCore.Infrastructure;
 
 namespace Microsoft.EntityFrameworkCore.Metadata.Internal
 {
