+++ src/EFCore/Extensions/Internal/EFPropertyExtensions.cs
@@ -30,27 +30,6 @@ namespace Microsoft.EntityFrameworkCore.Internal
         ///     any release. You should only use it directly in your code with extreme caution and knowing that
         ///     doing so can result in application failures when updating to a new Entity Framework Core release.
         /// </summary>
-        public static Expression CreateKeyAccessExpression(
-            [NotNull] this Expression target,
-            [NotNull] IReadOnlyList<IProperty> properties,
-            bool makeNullable = false)
-            => properties.Count == 1
-                ? target.CreateEFPropertyExpression(properties[0], makeNullable)
-                : Expression.New(
-                    AnonymousObject.AnonymousObjectCtor,
-                    Expression.NewArrayInit(
-                        typeof(object),
-                        properties
-                            .Select(p => Expression.Convert(target.CreateEFPropertyExpression(p, makeNullable), typeof(object)))
-                            .Cast<Expression>()
-                            .ToArray()));
-
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
         public static Expression CreateEFPropertyExpression(
             [NotNull] this Expression target,
             [NotNull] IPropertyBase property,
