+++ src/EFCore/Extensions/MutableEntityTypeExtensions.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
 using JetBrains.Annotations;
@@ -31,6 +32,16 @@ namespace Microsoft.EntityFrameworkCore
             => (IMutableEntityType)((IEntityType)entityType).GetRootType();
 
         /// <summary>
+        ///     Gets all types in the model which a given entity type derives from.
+        /// </summary>
+        /// <param name="entityType"> The type to find base types. </param>
+        /// <returns>
+        ///     The base types.
+        /// </returns>
+        public static IEnumerable<IMutableEntityType> GetAllBaseTypes([NotNull] this IMutableEntityType entityType)
+            => entityType.GetAllBaseTypesAscending().Reverse().Cast<IMutableEntityType>();
+
+        /// <summary>
         ///     Gets all types in the model that derive from a given entity type.
         /// </summary>
         /// <param name="entityType"> The base type to find types that derive from. </param>
@@ -53,6 +64,14 @@ namespace Microsoft.EntityFrameworkCore
         /// <returns> The derived types. </returns>
         public static IEnumerable<IMutableEntityType> GetDirectlyDerivedTypes([NotNull] this IMutableEntityType entityType)
             => ((EntityType)entityType).GetDirectlyDerivedTypes();
+
+        /// <summary>
+        ///     Returns all types in hierarchy of the given <see cref="IMutableEntityType" />.
+        /// </summary>
+        /// <param name="entityType"> The entity type. </param>
+        /// <returns> All types in the hierarchy. </returns>
+        public static IEnumerable<IMutableEntityType> GetTypesInHierarchy([NotNull] this IMutableEntityType entityType)
+            => entityType.GetAllBaseTypes().Concat(entityType.GetDerivedTypesInclusive()).Cast<IMutableEntityType>();
 
         /// <summary>
         ///     <para>
