+++ src/EFCore/Infrastructure/TypeExtensions.cs
@@ -2,6 +2,7 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
+using System.Linq;
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore.Internal;
 
@@ -35,5 +36,23 @@ namespace Microsoft.EntityFrameworkCore.Infrastructure
         /// <returns> The human-readable name. </returns>
         public static string ShortDisplayName([NotNull] this Type type)
             => type.DisplayName(fullName: false);
+
+        /// <summary>
+        ///     Gets a value indicating whether this type is same as or implements <see cref="IQueryable" />
+        /// </summary>
+        /// <param name="type"> The type to check. </param>
+        /// <returns>
+        ///     <c>True</c> if the type is same as or implements <see cref="IQueryable" />, otherwise <c>false</c>.
+        /// </returns>
+        public static bool IsQueryableType([NotNull] this Type type)
+        {
+            if (type.IsGenericType
+                && type.GetGenericTypeDefinition() == typeof(IQueryable<>))
+            {
+                return true;
+            }
+
+            return type.GetInterfaces().Any(i => i.IsGenericType && i.GetGenericTypeDefinition() == typeof(IQueryable<>));
+        }
     }
 }
