+++ src/EFCore/Metadata/Internal/PropertyBaseExtensions.cs
@@ -53,15 +53,6 @@ namespace Microsoft.EntityFrameworkCore.Metadata.Internal
         ///     any release. You should only use it directly in your code with extreme caution and knowing that
         ///     doing so can result in application failures when updating to a new Entity Framework Core release.
         /// </summary>
-        public static int GetIndex([NotNull] this IPropertyBase property)
-            => property.GetPropertyIndexes().Index;
-
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
         public static int GetOriginalValueIndex([NotNull] this IPropertyBase propertyBase)
             => propertyBase.GetPropertyIndexes().OriginalValueIndex;
